[feat] 프롬포트 작성 및 llm 호출 api 작성 + 테스트

+++ backend/src/main/java/com/example/mwo_do_shil/domain/recommand/dto/RectDto.java
@@ -1,17 +1,20 @@
 package com.example.mwo_do_shil.domain.recommand.dto;
 
+import io.swagger.v3.oas.annotations.Parameter;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 
 @Getter
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class RectDto {
-    private Double minX;
-    private Double minY;
-    private Double maxX;
-    private Double maxY;
+    private BigDecimal minX;
+    private BigDecimal minY;
+    private BigDecimal maxX;
+    private BigDecimal maxY;
 }
